requirements.txt
python-telegram-bot==20.8
yt-dlp
flask
requests
aiohttp
